Updating Changes

+++ src/main/java/com/satish/importJob/component/MutualFundScraper.java
@@ -13,6 +13,7 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 @Component
@@ -21,9 +22,10 @@ public class MutualFundScraper {
     @Autowired
     MutualFundPriceService mutualFundPriceService;
     @Scheduled(initialDelay = 0, fixedDelay = 86400000)
-    public void scrapeMutualFundData() throws InterruptedException {
+    public void scrapeMutualFundData()  {
         List<MutualFundPrice> mutualFundPriceList = new ArrayList<>();
         try {
+            logger.info("Mutual Fund Import Process Started");
             // URL of the ET Money mutual fund NAV page
             String url = "https://www.etmoney.com/mutual-funds/filter/latest-mutual-fund-nav";
 
@@ -37,17 +39,26 @@ public class MutualFundScraper {
                 Elements columns = fundElement.select("td");
                 if(columns.size()>5){
                     String fundName = columns.get(0).text();  // Adjust the selector as needed
-                    String nav = columns.get(2).text();   // Adjust the selector as needed
-                    System.out.println("Fund Name: " + fundName + ", NAV: " + nav);
-                    mutualFundPriceList.add(new MutualFundPrice(fundName,Double.valueOf(nav.substring(1,nav.length()))));
+                    String nav = columns.get(2).text();
+                    String navDate = columns.get(4).text();
+                    // Adjust the selector as needed
+                    //noinspection StringOperationCanBeSimplified
+                    mutualFundPriceList.add(new MutualFundPrice(fundName,Double.valueOf(nav.substring(1,nav.length())),
+                            navDate,new Date()));
                 }
 
             }
         } catch (Exception e) {
            logger.error("Error while scrapping Mutual Fund Data {}",e);
         }
-        mutualFundPriceService.deleteAll();
-        mutualFundPriceService.saveMutualFunds(mutualFundPriceList);
+        if(mutualFundPriceList.size()>0){
+            logger.info("Mutual Fund Import Process Completed");
+            mutualFundPriceService.deleteAll();
+            mutualFundPriceService.saveMutualFunds(mutualFundPriceList);
+        }else{
+            logger.warn("Mutual Fund Import is Having issues please verify it");
+        }
+
 
         }
