+++ src/main/java/com/satish/importJob/component/StockScraper.java
@@ -12,6 +12,7 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 @Component
@@ -23,8 +24,9 @@ public class StockScraper {
 
     // Define the cron expression for scheduling
     @Scheduled(initialDelay = 0, fixedDelay = 86400000)
-    public void scrapeStockData() throws InterruptedException {
+    public void scrapeStockData()  {
         List<StockPrice> stockPriceList = new ArrayList<>();
+        logger.info("Stock Price Import Started");
         for (int i = 1; i <= 132; i++) {
             String url = "https://www.screener.in/screens/356009/bse-nse-company-list/?page=" + i;
             try {
@@ -40,14 +42,12 @@ public class StockScraper {
                     if (columns.size() > 2) {// Ensure there are enough columns
                         String name = columns.get(1).text();  // Adjust index if necessary
                         String cmp = columns.get(2).text();   // Adjust index if necessary
-                        System.out.println("Name: " + name + ", CMP Rs: " + cmp);
                         if(!name.contains("Median")){
-                            stockPriceList.add(new StockPrice(name,Double.valueOf(cmp)));
+                            stockPriceList.add(new StockPrice(name,Double.valueOf(cmp),new Date()));
                         }
                         // You can save the data to MongoDB or any other storage here
                     }
                 }
-                System.out.println(url);
                 Thread.sleep(1000); // Sleep for 1 second to avoid hitting the server too frequently
             } catch (Exception e) {
                 logger.error("Error in importing stock prices : {}",e);
@@ -55,7 +55,13 @@ public class StockScraper {
 
 
         }
-        stockPriceService.deleteAll();
-        stockPriceService.saveStocks(stockPriceList);
+        if(stockPriceList.size()>0){
+            logger.info("Stock Price Import Completed");
+            stockPriceService.deleteAll();
+            stockPriceService.saveStocks(stockPriceList);
+        }else{
+            logger.warn("Issue with Stock Import please verify it");
+        }
+
     }
 }
